pyproject.toml
[tool.poetry]
name = "streamlit-demo"
version = "0.1.0"
description = ""
authors = ["chesterXalan <mastenfor26@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
streamlit-option-menu = "^0"
openai = "^1.35.13"
python-dotenv = "^1.0.1"
langchain-core = "^0"
langchain-community = "^0"
langchain-openai = "^0"


[tool.poetry.group.dev.dependencies]
ruff = "^0"
mypy = "^1.10.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"


[tool.poetry.group.jupyter.dependencies]
notebook = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# 以下都是靜態檢查設定

[tool.ruff]
target-version = "py312"
output-format = "concise"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    ### Intentionally disabled
    "E501", # line-too-long (format 會自動處理)
    "COM812",  # missing-trailing-comma  (會與 ruff format 衝突)
    "ISC001",  # ingle-line-implicit-string-concatenation (會與 ruff format 衝突)
    "UP040",  # non-pep695-type-alias  (mypy no supported)
    "S101", # assert  (for pytest)
    "ANN101",  # missing-type-self  (self 缺少類型註釋)
    "B008",  # unction-call-in-default-argument (不要在默認參數中調用函數)
    "FBT003",  # boolean-positional-value-in-call (不要使用 True/False 作為函數的位置參數)
    "RUF001",  # ambiguous-unicode-character-string  (不要使用不明確的 Unicode 字元)
    "RUF002",  # ambiguous-unicode-character-docstring  (不要使用不明確的 Unicode 字元)
    "RUF003",  # ambiguous-unicode-character-comment  (不要使用不明確的 Unicode 字元)
    "RUF005",  # collection-literal-concatenation  (不要使用集合文字串連)
    "BLE001",  # flake8-blind-except  (不要使用 BaseException 或 Exception)
    "TID252",  # relative-imports  (不要使用父母/兄弟姊妹級別以外的 相對 import)
    "SIM117",  # multiple-with-statements  (不要使用嵌套 with)
    "B904",  # raise-without-from-inside-except  (不要使用 raise 但沒有 from)
    "FIX002",  # flake8-fixme  (註釋包含 TODO)
    "TD",  # flake8-todos  (註釋包含 TODO 的一大堆限制)

    ### TODO: Enable gradually
    "EM",  # flake8-errmsg  (錯誤訊息字串要先分配給變數)
    "D",  # pydocstyle,  (docstring 檢查)
    "T",  # flake8-print  (不要使用 print)
    "PTH",  # flake8-use-pathlib  (不要使用 os.path，改用 pathlib)
    "ERA001",  # commented-out-code  (不要註解程式碼)
    "INP001",  # implicit-namespace-package  (缺少 __init__.py)
    "TRY003",  # raise-vanilla-args  (自定義例外包含額外的長訊息)
    "EXE002",  # shebang-missing-executable-file (尚不確定怎解決)
]

# <https://mypy.readthedocs.io/en/stable/config_file.html>
[tool.mypy]
python_version = "3.12"
strict = true
exclude = ["playground"]
follow_imports = "silent"
pretty = true
no_site_packages = true
ignore_missing_imports = true
disallow_subclassing_any = false
disallow_untyped_decorators = false
